refactor: extract vaadin bundles plugin into separate package
