+++ flow-server/src/main/resources/plugins/plugins.json
@@ -7,6 +7,7 @@
     "rollup-plugin-postcss-lit-custom",
     "react-function-location-plugin",
     "rollup-plugin-vaadin-i18n",
-    "vite-plugin-service-worker"
+    "vite-plugin-service-worker",
+    "vite-plugin-vaadin-bundles"
   ]
 }
